+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+* The `name` keyword argument to `submit_job` has been deprecated and replaced with `alias`. (#13)
+
 ## Version v0.1.2 - 2023-06-26
 
 ### Fixed